requirements.txt
python-dotenv==1.0.1
requests==2.32.3
jinja2==3.1.4
rapidfuzz==3.10.0
typer==0.12.5
rich==13.7.1
fastapi==0.112.2
uvicorn==0.30.6
pydantic==2.8.2
pyyaml==6.0.2
anyio==4.4.0
google-api-python-client==2.147.0
google-auth==2.33.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
stripe==9.13.0
email-validator==2.2.0
